Add variables to the `star` example
This PR discusses the geometry of star shapes, and provides variables that users can modify to control what kind of star they draw. It retains the classic five-pointed star as the default setting, but adds documentation explaining how to draw stars of other counts and sharpnesses by modifying the control variables.