Fixed unsoundness in AccessControl when processing RequiredTurtles::All
Scenario:

1. draw a line for a long time with a turtle that already exists
2. send a Clear request which then has to wait for the line
3. send CreateTurtle which runs immediately while clear is still waiting
4. begin drawing line with new turtle while Clear is still waiting

Unsound: if the first line finishes before the new turtle line is
done animating, clear will be given access to all turtles, including one
that it did not request access to when it was initially queued.

The reason for this is because we call turtle_ids() both before and after
waiting for the data to be ready. It's possible that the set of IDs may
have changed during that time. If we are not careful, we will accidentally
grant access to a turtle we did not request access to.

Solution: `All` should only provide access to the turtles available when
it was queued, not all of the turtles available when it is ready to run.

With this, the Clear command in the scenario above would not destroy the
line being drawn by the new turtle.

Alternative solution: `CreateTurtle` should wait until any request using
`RequiredTurtles::All` has completed. We could implement this by adding a
`DataChannel` specifically for requests with `RequiredTurtles::All`. The
`CreateTurtle` request would have to wait in that queue until no one else
is modifying "all" of the turltes at once. This would cause the line being
animated in step 4 of the scenario above to not begin drawing until *after*
the Clear command has finished.

In either solution, the Clear command always only removes the lines for
turtles that exist at the time when it is initially queued. This bug can
only be simulated using the request API directly. The interface of Drawing
and AsyncDrawing actually already prevent this situation from happening
because both add_turtle() and clear() take &mut self as their first
argument. Thus, it is impossible to run add_turtle() while clear() is
still waiting to be executed.

Given that both solutions effectively have the same behaviour in our
current public API, it was decided to implement the first one. The first
solution is easier to implement and is a smaller change.

Note that even if the public API doesn't allow for this unsoundness, it
was still important to fix it to avoid any future bugs that could be
introduced as this crate evolves.