changed to cfg target instead of feature for docs_image
I have cfg set to docs_image (I know I haven't renamed it yet to docs_images...)
so that it runs all the cfg(docs_image) code.
However it fails to import 'turtle_docs_helper' as it is not an external crate.

From reading up on this it should work... although it does suggest just
to use features for doing this kind of thing in the docs.

https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html#platform-specific-dependencies