Implement utilities for upcoming Ferris example
Began work on the Ferris example by defining the necessary constants,
functions and methods in order to have it easier later on. The plan is
to rely on cubic Bézier curves to avoid building lots of arcs by hand.
Note that in the current state, it does not compile.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>