Simplify forward movements for Ferris example
As I want to keep the drawing as independent as possible from the window
size, I added a few functions to transform the original segments and
Bézier curves to fit in the canvas. This commit adds another that simply
regroups the usual operation of setting the heading and going forwards.
It helps towards reducing the number of lines and complexity of the
example.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>